+++ crates/jsonschema-py/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Meta-schema validation support. [#442](https://github.com/Stranger6667/jsonschema/issues/442)
+
 ## [0.27.1] - 2024-12-24
 
 ### Added